+++ cloudfront.go
@@ -4,7 +4,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/cloudfront"
-	"log"
 )
 
 // CloudFrontProcessor holds the cloudfront-related actions